# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_root:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.UserPermission:
    properties:
      action:
        type: string
      created_at:
        type: string
      id:
        type: integer
      resource:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.UserResponse:
    properties:
      permissions:
        items:
          $ref: '#/definitions/model.UserPermission'
        type: array
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.UserUpdateRequest:
    properties:
      is_root:
        type: boolean
      status:
        enum:
        - active
        - inactive
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  util.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Private File Storage System API documentation
  title: PFSS API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get a list of users with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - Bearer: []
      summary: List users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get details of a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user details
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - Bearer: []
      summary: Update user
      tags:
      - users
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"